Update SymcDLPPerms15x.sh
10/24/2019
Fixed issues with string substitution in the MOD_PATH string variable passed to chmod that included a space in the path for Enforce 15.1. 

If the /opt/Symantec/DataLossPrevention/Enforce\ Server path itself has incorrect permissions, still throwing an error on chmod because the if statement to check the version depends on 15.1 to be present to use the substitution string. Still need to set a flag to check for perms on this one folder if the 15.1 flag is triggered so this file can also be modified.